Introduce 'minor' invocations
Resolving refs, etc adds noise to the logs. Simple invocations which
produce a single line of machine-significant output don't really need to
be console-logged, so provide a separate method for logging them.

The 'Type' property of the log entry struct is getting unwieldy. Move
formatting decisions to the call sites.