Add facility for overwriting stashes
The original design called for the remote stash to be append-only, but
we need the ability to overwrite stashes.

There is no simple way to have the server notify the client that its
stashes may be out of date. Instead, provide `--overwrite` arguments to
manually force the tool to re-download or -upload stashes.

refs INF-1149